Critical bugs exposed by Opteron experiments. includes premature freeing of the state machine message pair array in readdirplus which were somehow not revealed on x86!? Added lib64 to openssl.m4 so that we also look for these libraries in the 64 bit path prior to the 32 bit paths. If openssl was not present or not configured, make fskey_get return a NULL key and 0 keylength instead of a random pointer that will cause us to crash with a seg fault.. All test module.mk had to be augmented with -lssl -lcrypto to get them to build on opterons. Maximum number of directory entries (requested at 1 time) cannot be larger than maximum number of handles that we can handle per server list attr request. Dont know how I can make this a compile-time test... Fixed a few warnings in the test programs for Opterons. Upshot is that with this patch, all getdents/getdents64/getdents_plus/getdents64_plus work correctly on Opterons to the best of my knowledge on the sample workloads ran!!
SVN revision 5179

+++ test/posix/openg-mpi.c
@@ -181,8 +181,8 @@ int main(int argc, char *argv[])
 			exit(1);
 		}
 		end = Wtime();
-		printf("Rank %d openg on %s yielded length %d [Time %g msec]\n",
-				rank, fname, hb.handle_length, msec_diff(&end, &begin));
+		printf("Rank %d openg on %s yielded length %ld [Time %g msec]\n",
+				rank, fname, (unsigned long) hb.handle_length, msec_diff(&end, &begin));
 	}
 
 	/* Broadcast the handle buffer to everyone */
